+++ vault-plugins/flant_iam/backend/backend.go
@@ -1,4 +1,4 @@
-package mock
+package backend
 
 import (
 	"context"
